refactor(schema): extract fetchRemoteSchema into shared actions module
Pull and rollback subcommands previously imported fetchRemoteSchema from
push/actions.ts, creating an awkward sibling-to-sibling dependency. Move
the helper to a new schema/actions.ts at the command root, matching the
existing convention used by components/actions.ts and stories/actions.ts.

Fixes WDX-263
