+++ packages/cli/src/commands/schema/push/index.ts
@@ -13,7 +13,8 @@ import type { SchemaPushOptions } from './constants';
 import type { SchemaData } from '../types';
 import { loadSchema } from './load-schema';
 import { diffSchema } from './diff-schema';
-import { buildChangesetEntries, executePush, fetchRemoteSchema, formatDiffOutput } from './actions';
+import { fetchRemoteSchema } from '../actions';
+import { buildChangesetEntries, executePush, formatDiffOutput } from './actions';
 import { resolveFolderReferences } from './resolve-folders';
 import { saveChangeset } from './changeset';
 import { analyzeBreakingChanges } from './migrations/analyze';
