+++ packages/cli/src/commands/schema/rollback/index.ts
@@ -9,7 +9,7 @@ import { getUI } from '../../../utils/ui';
 import { session } from '../../../session';
 import { resolvePath } from '../../../utils/filesystem';
 import { schemaCommand } from '../command';
-import { fetchRemoteSchema } from '../push/actions';
+import { fetchRemoteSchema } from '../actions';
 import { saveChangeset } from '../push/changeset';
 import type { SchemaRollbackOptions } from './constants';
 import { buildRollbackOps, executeRollback, formatRollbackOutput, listChangesets, loadChangeset } from './actions';
